get rid of warnings

+++ csafety/src/c-objects.c
@@ -26,7 +26,7 @@ Presented below is a method of "programming with objects" in C.
 			return 1;
 
 		/* allocate memory and populate array members */
-		for (o=p; o-p < lim ; ++o)
+		for (o=p; (size_t)(o-p) < lim ; ++o)
 		{
 			if (!(o->rp = malloc(sizeof(char) * 100)))
 				return 1;
@@ -34,11 +34,11 @@ Presented below is a method of "programming with objects" in C.
 		} 
 
 		/* test the array of objects (fast write!) */
-		for (o=p; o-p < lim ; ++o)
+		for (o=p; (size_t)(o-p) < lim ; ++o)
 			write(1, o->rp, o->len);
 
 		/* free the objects' member memory allocations */
-		for (o=p; o-p < lim ; ++o)
+		for (o=p; (size_t)(o-p) < lim ; ++o)
 	                free(o->rp);
 
 		/* free the object */
